[REFACTOR] Final Adjustments

+++ src/main/java/br/ufpr/webII/trabalhoFinal/infra/service/EmployeeCrud.java
@@ -17,21 +17,33 @@ public class EmployeeCrud {
     @Autowired
     private DaoFactory daoFactory;
 
+    @Autowired
+    private TokenService tokenSrv;
+
     public void registerEmployee(EmployeeInputDTO data) {
         try {
             Employee employee = new Employee(data);
             EmployeeDao employeeSQLDao = daoFactory.getEmployeeDao();
             employeeSQLDao.insert(employee);
+        } catch (IllegalArgumentException e) {
+            throw e;
         } catch (Exception e){
             throw new RegisteringException("Erro ao inserir funcionário: " + e.getMessage(), e);
         }
     }
 
-    public void deleteEmployee(Long userId) {
+    public void deleteEmployee(Long userId, String auth) {
         try{
+            String authorization = tokenSrv.getAuthorizationToken(auth);
+            Long userIdToken = tokenSrv.getUserId(authorization);
+            if(userIdToken.equals(userId)){
+                throw new IllegalArgumentException("Não é possível excluir o próprio usuário");
+            }
             Employee employee = new Employee(userId);
             EmployeeDao employeeSQLDao = daoFactory.getEmployeeDao();
             employeeSQLDao.delete(employee);
+        } catch (IllegalArgumentException e){
+            throw e;
         } catch (Exception e){
             throw new RegisteringException("Erro ao excluir o empregado: " + e.getMessage(), e);
         }
